:seedling: Add dependency remediation in raw results instead of at log time
